Dataset Commit
In this commit, we upload the datasets that we will be using for this project. Each dataset has its unique properties related to nutritional information. All of this information is taken from the US Department of Agriculture (https://fdc.nal.usda.gov/). Therefore, it can be assured that the data is true. There are no false claims made.