Final Commit: - Project Submission
This is the final commit of the entire event. This includes adding enhancements to the code, a readme file to understand the problem statement, solution, roadmap, and code references. All of these codes combined give an efficient telegram nutrition bot that can answer almost all user queries and provide accurate nutritional information.